- VendingMachineImpl:  - Write addCurrency logic (WIP)  - Write getProducts logic (WIP)

+++ src/main/java/se/lexicon/data/IVendingMachine.java
@@ -3,7 +3,7 @@ package se.lexicon.data;
 import se.lexicon.model.Product;
 
 public interface IVendingMachine {
-    void addCurrency(double amount);
+    void addCurrency(int amount);
     int getBalance();
     Product request(int id);
     int endSession();
